requirement.txt
alembic==1.4.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cliff==3.1.0
cloudpickle==1.3.0
cmaes==0.3.2
cmd2==0.8.9
colorlog==4.1.0
configparser==5.0.0
databricks-cli==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker==4.2.0
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.2
GitPython==3.1.0
gorilla==0.3.0
gunicorn==20.0.4
hydra-core==0.11.3
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
jupytext==1.4.1
Mako==1.1.2
MarkupSafe==1.1.1
mistune==0.8.4
mlflow==1.7.2
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.4.5
notebook==6.0.3
numpy==1.18.2
omegaconf==1.4.1
optuna==1.3.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prettytable==0.7.2
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptvsd==4.3.2
ptyprocess==0.6.0
pudb==2019.2
Pygments==2.6.1
pyparsing==2.4.6
pyperclip==1.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
querystring-parser==1.2.4
requests==2.23.0
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.85
simplejson==3.17.0
six==1.14.0
smmap==3.0.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
stevedore==1.32.0
tabulate==0.8.7
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
urwid==2.1.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
